Update test files
The test files where not expecting the new additional information about the methods.

Also adapt the simple_struct.rs that the headlines of the documentation start with four '#' to avoid wrong headline structure inside the final document.

Also add a hint to README.md so that the user is aware of it.